# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Private API for ORD",
        "title": "Rocketbot ORD API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Michal Žídek",
            "url": "http://app.rocketbot.pro",
            "email": "m1chlcz18@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "89.116.25.234:7500",
    "basePath": "/api",
    "paths": {
        "/address": {
            "get": {
                "description": "Get new BTC Address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Daemon"
                ],
                "summary": "Get new BTC Address",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.NewAddressRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    }
                }
            }
        },
        "/estimate": {
            "post": {
                "description": "Estimate inscription cost !!!Don't use this method!!!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Estimate inscription cost !!!Don't use this method!!!",
                "parameters": [
                    {
                        "description": "Image URL from hosting service and number of blocks",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EstimateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Inscribe"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    }
                }
            }
        },
        "/feerate": {
            "get": {
                "description": "Get fee rate for transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fees"
                ],
                "summary": "Get fee rate for transaction",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FeeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    }
                }
            }
        },
        "/inscription/image": {
            "get": {
                "description": "Get Base64 image from Inscription ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Get Base64 image from Inscription ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Inscription ID",
                        "name": "idInscription",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.InscriptionPicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    }
                }
            }
        },
        "/inscriptions": {
            "get": {
                "description": "List of Inscriptions in the wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inscriptions"
                ],
                "summary": "List of Inscriptions in the wallet",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ListInscriptionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    }
                }
            }
        },
        "/mint": {
            "post": {
                "description": "Mint an Inscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Mint an Inscription",
                "parameters": [
                    {
                        "description": "File in base64 and file type",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MintRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Inscribe"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    }
                }
            }
        },
        "/nsfw": {
            "post": {
                "description": "Test picture for NSFW content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NSFW"
                ],
                "summary": "Test picture for NSFW content",
                "parameters": [
                    {
                        "description": "File in base64 and filename",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TestPicReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TestPicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    }
                }
            }
        },
        "/send": {
            "post": {
                "description": "Send an Inscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inscriptions"
                ],
                "summary": "Send an Inscription",
                "parameters": [
                    {
                        "description": "File in base64 and file type",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Inscribe"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorHTTP"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ErrorHTTP": {
            "type": "object",
            "properties": {
                "errorMessage": {
                    "type": "string"
                },
                "hasError": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.EstimateRequest": {
            "type": "object",
            "properties": {
                "blocks": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                }
            }
        },
        "models.FeeResponse": {
            "type": "object",
            "properties": {
                "feeRate": {
                    "type": "integer"
                },
                "hasError": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Inscribe": {
            "type": "object",
            "properties": {
                "commit": {
                    "type": "string"
                },
                "fees": {
                    "type": "integer"
                },
                "inscription": {
                    "type": "string"
                },
                "reveal": {
                    "type": "string"
                }
            }
        },
...
        "models.NewAddressRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
...
    }
}